# docs/static/openapi.yaml
openapi: 3.0.0
info:
  contact:
    name: AltMount
    url: https://github.com/javi11/altmount/issues
  description: REST API for AltMount — Usenet WebDAV server with NZB queue, health
    monitoring, ARR integration, FUSE/rclone mounting, and Stremio addon.
  license:
    name: MIT
  termsOfService: http://altmount.kipsilabs.top
  title: AltMount API
  version: "1.0"
paths:
...
  "/stremio/{key}/manifest.json":
    get:
      description: Returns the Stremio addon manifest JSON for installation. The key
        authenticates the addon.
      parameters:
        - description: Download key (SHA256 of API key)
          in: path
          name: key
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.stremioManifest"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      summary: Stremio addon manifest
      tags:
        - Stremio
  "/stremio/{key}/play":
    get:
      description: Downloads the NZB from Prowlarr by URL, queues it with high
        priority, waits for download completion, then redirects (302) to the
        first media stream URL.
      parameters:
        - description: Download key (SHA256 of API key)
          in: path
          name: key
          required: true
          schema:
            type: string
        - description: Prowlarr NZB download URL
          in: query
          name: url
          required: true
          schema:
            type: string
        - description: Safe filename title for the NZB
          in: query
          name: title
          schema:
            type: string
      responses:
        "302":
          description: Redirects to media stream URL
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      summary: Play Stremio NZB stream
      tags:
        - Stremio
  "/stremio/{key}/stream/{type}/{id}.json":
    get:
      description: Searches Prowlarr for matching NZBs and returns Stremio-compatible
        stream URL options.
      parameters:
        - description: Download key
          in: path
          name: key
          required: true
          schema:
            type: string
        - description: Content type (movie or series)
          in: path
          name: type
          required: true
          schema:
            type: string
        - description: Stremio content ID (e.g. tt1234567)
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      summary: Stremio stream handler
      tags:
        - Stremio
...
servers:
  - url: //localhost:8080/api
components:
...
  securitySchemes:
    ApiKeyAuth:
      in: query
      name: apikey
      type: apiKey
  schemas:
    api.APIError:
      properties:
        code:
          type: string
        details:
          type: string
        message:
          type: string
      type: object
    api.APIMeta:
      properties:
        count:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
      type: object
    api.APIResponse:
      properties:
        data: {}
        error:
          $ref: "#/components/schemas/api.APIError"
        meta:
          $ref: "#/components/schemas/api.APIMeta"
        success:
          type: boolean
      type: object
...
    api.stremioManifest:
      properties:
        catalogs:
          items: {}
          type: array
        description:
          type: string
        id:
          type: string
        idPrefixes:
          items:
            type: string
          type: array
        name:
          type: string
        resources:
          items:
            type: string
          type: array
        types:
          items:
            type: string
          type: array
        version:
          type: string
      type: object
...